+++ NEWS.md
@@ -2,6 +2,11 @@
 title: "NEWS"
 output: html_document
 ---
+## backbone 3.0.3
+
+* Updated/added generic `print()`, `summary()` and `plot()` functions for backbone objects
+* Corrected p-values in asymmetric marginal likelihood filter (h/t @florian-huber)
+
 ## backbone 3.0.2
 
 * minor updates to unit tests